Fix xattr detection on OpenBSD
OpenBSD doesn't have extended attributes
but we failed to detect that.

This is because the last xattr check, which checked for the Solaris
implementation, decided that we had xattr support when one of
the functions openat, fstatat, unlinkat, fchownat or futimesat was found.

The simple solution is to remove the Solaris check, since we don't implement
xattr support on Solaris anyway.

Fixes #312