Switch to non-recursive make
Recursive make clean doesn't work.
utest/ and src/ directories are on the same level
and utest/ needs sources from src/. Make clean tries
to deleted dependency files that have already been deleted and fails.

This is one of the issues with recursive make, so switch
to a single non-recursive Makefile. An added advantage is
that it's faster, since we don't build stuff twice.