Install the config files in $(sysconfdir)
Unfortunately we need to do some mangling of variables
first, to allow users to override them on the make command line.
The variables we substitute are runstatedir, sbindir, scriptdir
and sysconfdir.

I left server_address hardcoded to ::1 for now. Detecting if a host
supports ipv6 at make time doesn't make much sense, the package could
easily end up on a host that doesn't.