[AIX] ignore exceptfds in stats writing on AIX
On AIX, when writing the stats json files, an exception is present on
the writing file descriptor. We haven't found out why it happens,
everything looks fine up to the select() call. We suspect it has
something to do with setting the fd to non-blocking. The solution is
clearly not the best, but is the one with the least impact.