AIX: Complete initial support (#560)
* [AIX] Add endianess-handling for AIX (GCC)

AIX is always big endian ppc64. This patch adds the neccessary defines
when using GNU GCC by using GCC builtins. This has not been tested
with IBM xlC, which defines _AIX too, but doesn't define __GNUC__.

* [AIX] Prefix send_file and send_file_w with burp_

This sed-replaces send_file with burp_send_file because AIX defines
send_file in /usr/include/sys/socket.h, resulting in "already defined"
errors during compiling. For consistency, it also renames the _w variant.

* [AIX] utest: mask hashes requiring GNU-extensions

* [AIX] Add readme file with dependencies

This assumes GCC 4.9 and IBM AIX 7.1

* [AIX] Add a configure call to the readme

* Fix big endian handling in src/hexmap.c

In I1e45ab6014657f4a52df949c8ff246d5a7070e41, the preprocessor tests
for little endian twice.

* [AIX] Enhance documentation

* [AIX] ignore exceptfds in stats writing on AIX

On AIX, when writing the stats json files, an exception is present on
the writing file descriptor. We haven't found out why it happens,
everything looks fine up to the select() call. We suspect it has
something to do with setting the fd to non-blocking. The solution is
clearly not the best, but is the one with the least impact.