Fall back to legacy ncurses detection if pkg-config cannot find it.
This works around Travis XCode builds apparently being unable to locate
ncurses.pc. Ideally the XCode build process should be tweaked to make
sure that file *can* be found, seeing as pkg-config in general does
work; unfortunately neither I nor the main author of burp know Macs well
enough to implement this.

Note that the way the legacy check is invoked now, ncurses won't be
searched for if pkg-config (or to be precise the relevant m4 macro) is
not present at all. Hopefully this won't become a problem until this
workaround is no longer needed.