Use pkg-config to detect ncurses (#590)
* Use pkg-config to detect ncurses

Ncurses can be built split into libncurses proper and libtinfo. The old
check looked inside libncurses for a symbol which in case of a split
build appears in libtinfo, resulting in ncurses support in burp being
incorrectly disabled.

Seeing as burp already uses pkg-config to detect Check, use it. This
takes care of the split-library scenario and also makes it possible to
get rid of all the #ifdef HAVE_NCURSES_H ... #elif
HAVE_NCURSES_NCURSES_H ... #endif bits from the code - a simple #ifdef
HAVE_NCURSES ought to work.

* Fall back to legacy ncurses detection if pkg-config cannot find it.

This works around Travis XCode builds apparently being unable to locate
ncurses.pc. Ideally the XCode build process should be tweaked to make
sure that file *can* be found, seeing as pkg-config in general does
work; unfortunately neither I nor the main author of burp know Macs well
enough to implement this.

Note that the way the legacy check is invoked now, ncurses won't be
searched for if pkg-config (or to be precise the relevant m4 macro) is
not present at all. Hopefully this won't become a problem until this
workaround is no longer needed.