Reorder examples and group them by action
Some command line usage examples were listed in between other examples
with no apparent logical relationship. This is an attempt to make the
list easier to parse by keeping all examples together for a given
action. E.g. all the '-a r' examples, then all the '-a L' examples, and
so forth.