Minor changes suggested for the documentation (#731)
* Remove references to yajl.

Since the yajl code is included in the burp source tree, it is not
necessary to install a separate, external yajl package to build
burp. See commit d7327dbd256e3358d40005f45fd56cf9669efb57.

* Clarify usage of backup_tool_script on the burp server.

The documentation didn't make it clear that the script is only
functional on the burp server, and is meant to be used by a system
administrator with root access. It is of no use to a regular client
user.

* Remove unimplemented diff options.

The burp manpage showed two examples of the '-a d' diff action for
diffing a specific completed backup with the 'next' backup to be made,
along with a note "DIFF OPTIONS NOT FULLY IMPLEMENTED YET". Including
these in the documentation served no purpose and could be confusing to
users.

* Save deleted documentation

Some diff options that were initially documented in the burp manpage but
later removed because they are not implemented, are saved here for
future reference if needed. This may not be the best place to put them,
or they may not need to be saved at all.

* Reorder examples and group them by action

Some command line usage examples were listed in between other examples
with no apparent logical relationship. This is an attempt to make the
list easier to parse by keeping all examples together for a given
action. E.g. all the '-a r' examples, then all the '-a L' examples, and
so forth.