Ensure O_NOATIME is passed to system calls (#772)
In the C programming language the "|" operator has a higher precedence
than the "?:" operator, see [1]. The expression

  O_RDONLY|O_BINARY|atime?0:O_NOATIME

used in "bfile_open" can be parenthesised as

  O_RDONLY|(O_BINARY|atime)?0:O_NOATIME

On Unix systems this happens to work as intended because Burp defines
"O_BINARY" as zero (0). The resulting system call with the "atime"
configuration option set to 0:

  open("/a/file", O_RDONLY|O_NOATIME) = 5

The case of "entries_in_directory" is similar, but in this case
"O_DIRECTORY" doesn't make it to the system call and neither does
"O_NOATIME":

  open("/a/directory", O_RDONLY) = 5

As a consequence atime is always updated on directories (unless
precluded by mount options). After adding parenthesis to the
expressions:

  open("/a/file", O_RDONLY|O_NOATIME) = 5
  open("/a/directory", O_RDONLY|O_DIRECTORY|O_NOATIME) = 5

[1] https://en.cppreference.com/w/c/language/operator_precedence