client: Don't rewrite config if not neccessary.
The client rewrites the config file after having the certificate signed
to update the "ssl_peer_cn" config option with the value received from
the server. Burp creates a temporary file next to the actual config file
in preparation for an atomic move should the files differ in content.
This patch modifies this behaviour in that it does not attempt to write
the temporary file in the first place if the "ssl_peer_cn" did not
change.

The original behaviour requires write permissions for the burp client
user to the directory where the main config file resides (usually
`/etc/burp`). If burp doesn't run as root, the entire config directory
would have to be writable for that user until the certificate signing
 has been completed successfully, even if the config file was correct
 from the beginning.