Fix VSS stripping behavior when headers span >1 network block.
When VSS header data straddles multiple network blocks (i.e. got <
bsidsize in bfile_write_vss_strip()), the pointer arithmatic when
memcpy()ing the next block is incorrect.

Since sid is a pointer to a structure, each addition by 1 increases
the pointer address by sizeof(*sid); therefore, when sidlen > 0, the
memcpy() unintentionally overwrites part of the struct BFILE it lives
within or the memory situated after it.

Casting sid to a char* in the memcpy() causes the addition to advance
the pointer by single bytes as intended.